Remove Interaction Property Methods in Interaction Model
