+++ src/abaqus/Interaction/InteractionModel.py
@@ -3,34 +3,25 @@ import typing
 from abaqusConstants import *
 from .InteractionContactControlModel import InteractionContactControlModel
 from .InteractionContactInitializationModel import InteractionContactInitializationModel
+from .InteractionContactStabilizationModel import InteractionContactStabilizationModel
 from .InteractionPropertyModel import InteractionPropertyModel
 from ..BasicGeometry.ModelDot import ModelDot
 from ..Datum.DatumAxis import DatumAxis
 from ..Interaction.AcousticImpedance import AcousticImpedance
-from ..Interaction.AcousticImpedanceProp import AcousticImpedanceProp
 from ..Interaction.ActuatorSensor import ActuatorSensor
-from ..Interaction.ActuatorSensorProp import ActuatorSensorProp
 from ..Interaction.CavityRadiation import CavityRadiation
-from ..Interaction.CavityRadiationProp import CavityRadiationProp
 from ..Interaction.ConcentratedFilmCondition import ConcentratedFilmCondition
 from ..Interaction.ConcentratedRadiationToAmbient import ConcentratedRadiationToAmbient
 from ..Interaction.ContactExp import ContactExp
-from ..Interaction.ContactProperty import ContactProperty
 from ..Interaction.ContactPropertyAssignment import ContactPropertyAssignment
 from ..Interaction.ContactStd import ContactStd
 from ..Interaction.CyclicSymmetry import CyclicSymmetry
 from ..Interaction.ElasticFoundation import ElasticFoundation
-from ..Interaction.ExpContactControl import ExpContactControl
 from ..Interaction.FilmCondition import FilmCondition
-from ..Interaction.FilmConditionProp import FilmConditionProp
 from ..Interaction.FluidCavity import FluidCavity
-from ..Interaction.FluidCavityProperty import FluidCavityProperty
 from ..Interaction.FluidExchange import FluidExchange
-from ..Interaction.FluidExchangeProperty import FluidExchangeProperty
 from ..Interaction.FluidInflator import FluidInflator
-from ..Interaction.FluidInflatorProperty import FluidInflatorProperty
 from ..Interaction.IncidentWave import IncidentWave
-from ..Interaction.IncidentWaveProperty import IncidentWaveProperty
 from ..Interaction.InitializationAssignment import InitializationAssignment
 from ..Interaction.MainSecondaryAssignment import MainSecondaryAssignment
 from ..Interaction.ModelChange import ModelChange
@@ -42,9 +33,6 @@ from ..Interaction.SelfContactStd import SelfContactStd
 from ..Interaction.SlidingTransitionAssignment import SlidingTransitionAssignment
 from ..Interaction.SmoothingAssignment import SmoothingAssignment
 from ..Interaction.StabilizationAssignment import StabilizationAssignment
-from ..Interaction.StdContactControl import StdContactControl
-from ..Interaction.StdInitialization import StdInitialization
-from ..Interaction.StdStabilization import StdStabilization
 from ..Interaction.StdXplCosimulation import StdXplCosimulation
 from ..Interaction.SurfaceFeatureAssignment import SurfaceFeatureAssignment
 from ..Interaction.SurfaceOffsetAssignment import SurfaceOffsetAssignment
@@ -59,6 +47,7 @@ from ..Region.RegionArray import RegionArray
 class InteractionModel(
     InteractionContactControlModel,
     InteractionContactInitializationModel,
+    InteractionContactStabilizationModel,
     InteractionPropertyModel,
 ):
     """Abaqus creates a Model object named `Model-1` when a session is started.
@@ -285,50 +274,6 @@ class InteractionModel(
         )
         return interaction
 
-    def AcousticImpedanceProp(
-        self,
-        name: str,
-        tableType: SymbolicConstant,
-        table: tuple,
-        frequencyDependency: Boolean = OFF,
-    ) -> AcousticImpedanceProp:
-        """This method creates an AcousticImpedanceProp object.
-
-        .. note:: 
-            This function can be accessed by:
-
-            .. code-block:: python
-
-                mdb.models[name].AcousticImpedanceProp
-
-        Parameters
-        ----------
-        name
-            A String specifying the interaction property repository key.
-        tableType
-            A SymbolicConstant specifying the type of tabular data to be defined. Possible values
-            are IMPEDANCE and ADMITTANCE.
-        table
-            A sequence of sequences of Floats specifying acoustic impedance properties.If
-            **tableType** = IMPEDANCE, each sequence of the table data specifies:The real part of the
-            complex impedance.The imaginary part of the complex impedance.Frequency, if the data
-            depend on frequency.If **tableType** = ADMITTANCE, each sequence of the table data
-            specifies:The real part of the complex admittance.The imaginary part of the complex
-            admittance.Frequency, if the data depend on frequency.
-        frequencyDependency
-            A Boolean specifying whether the **table** data depend on frequency. The default value is
-            OFF.
-
-        Returns
-        -------
-        AcousticImpedanceProp
-            An :py:class:`~abaqus.Interaction.AcousticImpedanceProp.AcousticImpedanceProp` object.
-        """
-        self.interactions[name] = interaction = AcousticImpedanceProp(
-            name, tableType, table, frequencyDependency
-        )
-        return interaction
-
     def ActuatorSensor(
         self,
         name: str,
@@ -393,39 +338,6 @@ class InteractionModel(
             userSubUel,
             dof,
             solutionDepVars,
-        )
-        return interaction
-
-    def ActuatorSensorProp(
-        self, name: str, realProperties: tuple = (), integerProperties: tuple = ()
-    ) -> ActuatorSensorProp:
-        """This method creates an ActuatorSensorProp object.
-
-        .. note:: 
-            This function can be accessed by:
-
-            .. code-block:: python
-
-                mdb.models[name].ActuatorSensorProp
-
-        Parameters
-        ----------
-        name
-            A String specifying the interaction property repository key.
-        realProperties
-            A sequence of Floats specifying the PROPS array used by user subroutine UEL. The default
-            value is an empty sequence.
-        integerProperties
-            A sequence of Ints specifying the JPROPS array used by user subroutine UEL. The default
-            value is an empty sequence.
-
-        Returns
-        -------
-        ActuatorSensorProp
-            An :py:class:`~abaqus.Interaction.ActuatorSensorProp.ActuatorSensorProp` object.
-        """
-        self.interactions[name] = interaction = ActuatorSensorProp(
-            name, realProperties, integerProperties
         )
         return interaction
 
@@ -679,45 +591,6 @@ class InteractionModel(
             reflectionSymPlane_2,
             reflectionSymPlane_3,
             reflectionSymZ,
-        )
-        return interaction
-
-    def CavityRadiationProp(
-        self,
-        name: str,
-        temperatureDependency: Boolean = OFF,
-        dependencies: int = 0,
-        property: tuple = (),
-    ) -> CavityRadiationProp:
-        """This method creates a CavityRadiationProp object.
-
-        .. note:: 
-            This function can be accessed by:
-
-            .. code-block:: python
-
-                mdb.models[name].CavityRadiationProp
-
-        Parameters
-        ----------
-        name
-            A String specifying the interaction property repository key.
-        temperatureDependency
-            A Boolean specifying whether the data depend on temperature. The default value is OFF.
-        dependencies
-            An Int specifying the number of field variable dependencies. The default value is 0.
-        property
-            A sequence of sequences of Floats specifying the following:The emissivity,
-            ϵϵ.Temperature, if the data depend on temperature.Value of the first field variable, if
-            the data depend on field variables.Value of the second field variable.Etc.
-
-        Returns
-        -------
-        CavityRadiationProp
-            A :py:class:`~abaqus.Interaction.CavityRadiationProp.CavityRadiationProp` object.
-        """
-        self.interactions[name] = interaction = CavityRadiationProp(
-            name, temperatureDependency, dependencies, property
         )
         return interaction
 
@@ -986,31 +859,6 @@ class InteractionModel(
         )
         return interaction
 
-    def ContactProperty(self, name: str) -> ContactProperty:
-        """This method creates a ContactProperty object.
-
-        .. note:: 
-            This function can be accessed by:
-
-            .. code-block:: python
-
-                mdb.models[name].ContactProperty
-
-        Parameters
-        ----------
-        name
-            A String specifying the interaction property repository key.
-
-        Returns
-        -------
-        ContactProperty
-            A :py:class:`~abaqus.Interaction.ContactProperty.ContactProperty` object.
-        """
-        interactionProperties = ContactProperty(name)
-        self.interactions[name] = interactionProperties
-        self.interactionProperties[name] = interactionProperties
-        return interactionProperties
-
     def ContactStd(
         self,
         name: str,
@@ -1240,71 +1088,6 @@ class InteractionModel(
         )
         return interaction
 
-    def ExpContactControl(
-        self,
-        name: str,
-        globTrkChoice: SymbolicConstant = DEFAULT,
-        globTrkInc: int = None,
-        fastLocalTrk: Boolean = ON,
-        scalePenalty: float = 1,
-        warpCheckPeriod: int = 20,
-        warpCutoff: float = 20,
-    ) -> ExpContactControl:
-        """This method creates an ExpContactControl object.
-
-        .. note:: 
-            This function can be accessed by:
-
-            .. code-block:: python
-
-                mdb.models[name].ExpContactControl
-
-        Parameters
-        ----------
-        name
-            A String specifying the contact controls repository key.
-        globTrkChoice
-            A SymbolicConstant specifying whether or not the default value will be used for the
-            maximum number of increments between global contact searches. Possible values are
-            DEFAULT and SPECIFY. The default value is DEFAULT.
-        globTrkInc
-            An Int specifying the maximum number of increments between global contact searches. The
-            **globTrkInc** argument applies only when **globTrkChoice** = SPECIFY. The default value is
-            100 for surface-to-surface contact and 4 for self-contact.
-        fastLocalTrk
-            A Boolean specifying whether to use the more computationally efficient local tracking
-            method. The default value is ON.
-        scalePenalty
-            A Float specifying the factor by which Abaqus/Explicit will scale the default penalty
-            stiffness to obtain the stiffnesses used for the penalty contact pairs. The default
-            value is 1.0.
-        warpCheckPeriod
-            An Int specifying the number of increments between checks for highly warped facets on
-            main surfaces. The default value is 20.
-        warpCutoff
-            A Float specifying the out-of-plane warping angle (in degrees), at which a facet will be
-            considered to be highly warped. The default value is 20.0.
-
-        Returns
-        -------
-        ExpContactControl
-            An :py:class:`~abaqus.Interaction.ExpContactControl.ExpContactControl` object.
-
-        Raises
-        ------
-        RangeError
-        """
-        self.interactions[name] = interaction = ExpContactControl(
-            name,
-            globTrkChoice,
-            globTrkInc,
-            fastLocalTrk,
-            scalePenalty,
-            warpCheckPeriod,
-            warpCutoff,
-        )
-        return interaction
-
     def FilmCondition(
         self,
         name: str,
@@ -1399,49 +1182,6 @@ class InteractionModel(
         )
         return interaction
 
-    def FilmConditionProp(
-        self,
-        name: str,
-        temperatureDependency: Boolean = OFF,
-        dependencies: int = 0,
-        property: tuple = (),
-    ) -> FilmConditionProp:
-        """This method creates a FilmConditionProp object.
-
-        .. note:: 
-            This function can be accessed by:
-
-            .. code-block:: python
-
-                mdb.models[name].FilmConditionProp
-
-        Parameters
-        ----------
-        name
-            A String specifying the interaction property repository key.
-        temperatureDependency
-            A Boolean specifying whether the data depend on temperature. The default value is OFF.
-        dependencies
-            An Int specifying the number of field variable dependencies. The default value is 0.
-        property
-            A sequence of sequences of Floats specifying the following:
-            
-            - The film coefficient, hh.
-            - Temperature, if the data depend on temperature.
-            - Value of the first field variable, if the data depend on field variables.
-            - Value of the second field variable.
-            - Etc.
-
-        Returns
-        -------
-        FilmConditionProp
-            A :py:class:`~abaqus.Interaction.FilmConditionProp.FilmConditionProp` object.
-        """
-        self.interactions[name] = interaction = FilmConditionProp(
-            name, temperatureDependency, dependencies, property
-        )
-        return interaction
-
     def FluidCavity(
         self,
         name: str,
@@ -1506,161 +1246,6 @@ class InteractionModel(
         )
         return interaction
 
-    def FluidCavityProperty(
-        self,
-        name: str,
-        definition: SymbolicConstant = HYDRAULIC,
-        fluidDensity: float = None,
-        molecularWeight: float = None,
-        useExpansion: Boolean = OFF,
-        expansionTempDep: Boolean = OFF,
-        expansionDependencies: int = 0,
-        referenceTemperature: float = 0,
-        expansionTable: tuple = (),
-        useBulkModulus: Boolean = OFF,
-        bulkModulusTempDep: Boolean = OFF,
-        bulkModulusDependencies: int = 0,
-        bulkModulusTable: tuple = (),
-        useCapacity: Boolean = OFF,
-        capacityType: SymbolicConstant = POLYNOMIAL,
-        capacityTempDep: Boolean = OFF,
-        capacityDependencies: int = 0,
-        capacityTable: tuple = (),
-    ) -> FluidCavityProperty:
-        """This method creates a FluidCavityProperty object.
-
-        .. note:: 
-            This function can be accessed by:
-
-            .. code-block:: python
-
-                mdb.models[name].FluidCavityProperty
-
-        Parameters
-        ----------
-        name
-            A String specifying the interaction property repository key.
-        definition
-            A SymbolicConstant specifying the type of fluid cavity property to be defined. Possible
-            values are HYDRAULIC and PNEUMATIC. The default value is HYDRAULIC.
-        fluidDensity
-            None or a Float specifying the reference fluid density. This argument is applicable only
-            when **definition** = HYDRAULIC, and is required in that case. The default value is None.
-        molecularWeight
-            None or a Float specifying the molecular weight of the ideal gas species. This argument
-            is applicable only when **definition** = PNEUMATIC, and is required in that case. The
-            default value is None.
-        useExpansion
-            A Boolean specifying whether thermal expansion coefficients will be defined. This
-            argument is applicable only when **definition** = HYDRAULIC. The default value is OFF.
-        expansionTempDep
-            A Boolean specifying whether the thermal fluid expansion data will have temperature
-            dependency. This argument is applicable only when **definition** = HYDRAULIC and when
-            **useExpansion** = True. The default value is OFF.
-        expansionDependencies
-            An Int specifying the number of field variable dependencies in the thermal fluid
-            expansion data. This argument is applicable only when **definition** = HYDRAULIC and when
-            **useExpansion** = True. The default value is 0.
-        referenceTemperature
-            A Float specifying the reference temperature for the coefficient of thermal expansion.
-            This argument is applicable only when **definition** = HYDRAULIC, when **useExpansion** = True,
-            and when either **expansionTempDep** = True or when **expansionDependencies** is greater than
-            0. The default value is 0.0.
-        expansionTable
-            A sequence of sequences of Floats specifying the thermal expansion coefficients. This
-            argument is applicable only when **definition** = HYDRAULIC and when **useExpansion** = True.
-            Each sequence contains the following data:
-            
-            - The mean coefficient of thermal expansion.
-            - Temperature, if the data depend on temperature.
-            - Value of the first field variable, if the data depend on field variables.
-            - Value of the second field variable.
-            - Etc.
-        useBulkModulus
-            A Boolean specifying whether fluid bulk modulus values will be defined. This argument is
-            applicable only when **definition** = HYDRAULIC. The default value is OFF.
-        bulkModulusTempDep
-            A Boolean specifying whether the fluid bulk modulus data will have temperature
-            dependency. This argument is applicable only when **definition** = HYDRAULIC and when
-            **useBulkModulus** = True. The default value is OFF.
-        bulkModulusDependencies
-            An Int specifying the number of field variable dependencies in the fluid bulk modulus
-            data. This argument is applicable only when **definition** = HYDRAULIC and when
-            **useBulkModulus** = True. The default value is 0.
-        bulkModulusTable
-            A sequence of sequences of Floats specifying the fluid bulk modulus values. This
-            argument is applicable only when **definition** = HYDRAULIC and when **useBulkModulus** = True.
-            Each sequence contains the following data:
-            
-            - The fluid bulk modulus.
-            - Temperature, if the data depend on temperature.
-            - Value of the first field variable, if the data depend on field variables.
-            - Value of the second field variable.
-            - Etc.
-        useCapacity
-            A Boolean specifying whether molar heat capacity values will be defined. This argument
-            is applicable only when **definition** = PNEUMATIC. The default value is OFF.
-        capacityType
-            A SymbolicConstant specifying the method to define the molar heat capacity. Possible
-            values are POLYNOMIAL and TABULAR. The default value is POLYNOMIAL.
-        capacityTempDep
-            A Boolean specifying whether the molar heat capacity data will have temperature
-            dependency. This argument is applicable only when **definition** = PNEUMATIC, when
-            **useCapacity** = True, and when **capacityType** = TABULAR. The default value is OFF.
-        capacityDependencies
-            An Int specifying the number of field variable dependencies in the molar heat capacity
-            data. This argument is applicable only when **definition** = PNEUMATIC, when
-            **useCapacity** = True, and when **capacityType** = TABULAR. The default value is 0.
-        capacityTable
-            A sequence of sequences of Floats specifying the molar heat capacity values in the form
-            of a polynomial expression. This argument is applicable only when
-            **definition** = PNEUMATIC, when **useCapacity** = True, and when **capacityType** = POLYNOMIAL. In
-            this form, only one sequence is specified and that sequence contains the following data:
-            
-            - The first molar heat capacity coefficient.
-            - The second molar heat capacity coefficient.
-            - The third molar heat capacity coefficient.
-            - The fourth molar heat capacity coefficient.
-            - The fifth molar heat capacity coefficient.
-              
-            Alternatively, the sequence data may specify the molar heat capacity values at constant
-            pressure for an ideal gas species. This argument is applicable only when
-            **definition** = PNEUMATIC, when **useCapacity** = True, and when **capacityType** = TABULAR. Each
-            sequence contains the following data:
-            
-            - The molar heat capacity at constant pressure.
-            - Temperature, if the data depend on temperature.
-            - Value of the first field variable, if the data depend on field variables.
-            - Value of the second field variable.
-            - Etc.
-
-        Returns
-        -------
-        FluidCavityProperty
-            A :py:class:`~abaqus.Interaction.FluidCavityProperty.FluidCavityProperty` object.
-        """
-        self.interactions[name] = interaction = FluidCavityProperty(
-            name,
-            definition,
-            fluidDensity,
-            molecularWeight,
-            useExpansion,
-            expansionTempDep,
-            expansionDependencies,
-            referenceTemperature,
-            expansionTable,
-            useBulkModulus,
-            bulkModulusTempDep,
-            bulkModulusDependencies,
-            bulkModulusTable,
-            useCapacity,
-            capacityType,
-            capacityTempDep,
-            capacityDependencies,
-            capacityTable,
-        )
-        return interaction
-
     def FluidExchange(
         self,
         name: str,
@@ -1716,109 +1301,6 @@ class InteractionModel(
         )
         return interaction
 
-    def FluidExchangeProperty(
-        self,
-        name: str,
-        dataTable: tuple,
-        definition: SymbolicConstant = BULK_VISCOSITY,
-        pressureDependency: Boolean = OFF,
-        temperatureDependency: Boolean = OFF,
-        fieldDependencies: int = 0,
-    ) -> FluidExchangeProperty:
-        """This method creates a FluidExchangeProperty object.
-
-        .. note:: 
-            This function can be accessed by:
-
-            .. code-block:: python
-
-                mdb.models[name].FluidExchangeProperty
-
-        Parameters
-        ----------
-        name
-            A String specifying the interaction property repository key.
-        dataTable
-            A sequence of sequences of Floats specifying the viscous and hydrodynamic resistance
-            coefficients when **definition** = BULK_VISCOSITY. Each sequence contains the following
-            data:
-            
-            - The viscous resistance coefficient.
-            - The hydrodynamic resistance coefficient.
-            - The average absolute pressure, if the data depend on pressure.
-            - The average temperature, if the data depend on temperature.
-            - The value of the first field variable, if the data depend on field variables.
-            - The value of the second field variable.
-            - Etc.
-            
-            Alternatively, the sequence data may specify the mass flow rate. This is applicable only
-            when **definition** = MASS_FLUX. In this form, only one sequence is specified and that
-            sequence contains the following data:
-            
-            - The mass flow rate per unit area.
-            
-            Alternatively, the sequence data may specify the mass rate leakage. This is applicable
-            only when **definition** = MASS_RATE_LEAK. Each sequence contains the following data:
-              
-            - The absolute value of the mass flow rate per unit area. (The first tabular value
-              entered must always be zero.)
-            - The absolute value of the pressure difference. (The first tabular value entered must
-              always be zero.)
-            - The average absolute pressure, if the data depend on pressure.
-            - The average temperature, if the data depend on temperature.
-            - The value of the first field variable, if the data depend on field variables.
-            - The value of the second field variable.
-            - Etc.
-            
-            Alternatively, the sequence data may specify the volume flow rate. This is applicable
-            only when **definition** = VOL_FLUX. In this form, only one sequence is specified and that
-            sequence contains the following data:
-            
-            - The volumetric flow rate per unit area.
-              Alternatively, the sequence data may specify the volume rate leakage. This is applicable
-              only when **definition** = VOL_RATE_LEAK. Each sequence contains the following data:
-            
-            - The absolute value of the volumetric flow rate per unit area. (The first tabular value
-              entered must always be zero.)
-            - The absolute value of the pressure difference. (The first tabular value entered must
-              always be zero.)
-            - The average absolute pressure, if the data depend on pressure.
-            - The average temperature, if the data depend on temperature.
-            - The value of the first field variable, if the data depend on field variables.
-            - The value of the second field variable.
-            - Etc.
-        definition
-            A SymbolicConstant specifying the type of fluid exchange property to be defined.
-            Possible values are BULK_VISCOSITY, MASS_FLUX, MASS_RATE_LEAK, VOL_FLUX, and
-            VOL_RATE_LEAK. The default value is BULK_VISCOSITY.
-        pressureDependency
-            A Boolean specifying whether the data will have pressure dependency. This argument is
-            applicable only when **definition** = BULK_VISCOSITY, or when **definition** = MASS_RATE_LEAK,
-            or when **definition** = VOL_RATE_LEAK. The default value is OFF.
-        temperatureDependency
-            A Boolean specifying whether the data will have temperature dependency. This argument is
-            applicable only when **definition** = BULK_VISCOSITY, or when **definition** = MASS_RATE_LEAK,
-            or when **definition** = VOL_RATE_LEAK. The default value is OFF.
-        fieldDependencies
-            An Int specifying the number of field variable dependencies in the data. This argument
-            is applicable only when **definition** = BULK_VISCOSITY, or when
-            **definition** = MASS_RATE_LEAK, or when **definition** = VOL_RATE_LEAK. The default value is 0.
-
-        Returns
-        -------
-        FluidExchangeProperty
-            A :py:class:`~abaqus.Interaction.FluidExchangeProperty.FluidExchangeProperty` object.
-        """
-        self.interactions[name] = interaction = FluidExchangeProperty(
-            name,
-            dataTable,
-            definition,
-            pressureDependency,
-            temperatureDependency,
-            fieldDependencies,
-        )
-        return interaction
-
     def FluidInflator(
         self,
         name: str,
@@ -1865,82 +1347,6 @@ class InteractionModel(
             interactionProperty,
             inflationTimeAmplitude,
             massFlowAmplitude,
-        )
-        return interaction
-
-    def FluidInflatorProperty(
-        self,
-        name: str,
-        definition: str,
-        effectiveArea: float,
-        tankVolume: float,
-        dischargeCoefficient: float = None,
-        dataTable: tuple = (),
-        numFluids: int = None,
-        mixtureType: str = "",
-        inflationTime: tuple = (),
-        fluidbehaviorName: tuple = (),
-        massFraction: tuple = (),
-    ) -> FluidInflatorProperty:
-        """This method creates a FluidInflatorProperty object.
-
-        .. note::
-            This function can be accessed by:
-
-            .. code-block:: python
-
-                mdb.models[name].FluidInflatorProperty
-
-        Parameters
-        ----------
-        name
-            A String specifying the interaction property repository key.
-        definition
-            A Symbolic constant specifying the method used for modeling the flow characteristics of
-            inflators. The default value is *definition*=DUAL PRESSURE. The possible values are DUAL
-            PRESSURE, PRESSURE AND MASS, TANK TEST, and TEMPERATURE AND MASS.
-        effectiveArea
-            A Float specifying the total inflator orifice area. This argument is applicable only if
-            *definition*=DUAL PRESSURE or *definition*=PRESSURE AND MASS.
-        tankVolume
-            A Float specifying the tank volume. This argument is applicable only if
-            *definition*=DUAL PRESSURE or *definition*=TANK TEST.
-        dischargeCoefficient
-            A Float specifying the discharge coefficient. This argument is applicable only if
-            *definition*=DUAL PRESSURE or *definition*=PRESSURE AND MASS.
-        dataTable
-            A sequence of sequences of Floats specifying the items described in the "Table data"
-            section below.
-        numFluids
-            An Int specifying the number of gas species used for this inflator.
-        mixtureType
-            A Symbolic constant specifying whether to use mass fraction or the molar fraction for a
-            mixture of ideal gases. The default value is MASS FRACTION. The possible values are MASS
-            FRACTION or MOLAR FRACTION.
-        inflationTime
-            A sequence of sequences of Floats specifying the inflation time.
-        fluidbehaviorName
-            A sequence of sequences of Strings specifying fluid behavior names.
-        massFraction
-            A sequence of sequences of Floats specifying the mass fraction or the molar fraction
-            corresponding to entered fluid behavior.
-
-        Returns
-        -------
-            A FluidInflatorProperty object.
-        """
-        self.interactions[name] = interaction = FluidInflatorProperty(
-            name,
-            definition,
-            effectiveArea,
-            tankVolume,
-            dischargeCoefficient,
-            dataTable,
-            numFluids,
-            mixtureType,
-            inflationTime,
-            fluidbehaviorName,
-            massFraction,
         )
         return interaction
 
@@ -2040,211 +1446,6 @@ class InteractionModel(
             referenceMagnitude,
             detonationTime,
             magnitudeFactor,
-        )
-        return interaction
-
-    def IncidentWaveProperty(
-        self,
-        name: str,
-        definition: SymbolicConstant = PLANAR,
-        propagationModel: SymbolicConstant = ACOUSTIC,
-        soundSpeed: float = None,
-        fluidDensity: float = None,
-        specificHeatRatio: float = None,
-        gravity: float = None,
-        atmosphericPressure: float = None,
-        dragCoefficient: float = None,
-        dragExponent: float = 2,
-        waveEffects: Boolean = ON,
-        chargeDensity: float = None,
-        chargeMass: float = None,
-        constantK1: float = None,
-        constantK2: float = None,
-        constantA: float = None,
-        constantB: float = None,
-        constantKc: float = None,
-        duration: float = None,
-        maximumSteps: int = 1500,
-        relativeStepControl: float = None,
-        absoluteStepControl: float = None,
-        stepControlExponent: float = 0,
-        genDecayA: float = 0,
-        genDecayB: float = 0,
-        genDecayC: float = 0,
-        seedNumber: int = None,
-        massTNT: float = None,
-        massFactor: float = 1,
-        lengthFactor: float = 1,
-        timeFactor: float = 1,
-        pressureFactor: float = 1,
-    ) -> IncidentWaveProperty:
-        """This method creates an IncidentWaveProperty object.
-
-        .. note:: 
-            This function can be accessed by:
-
-            .. code-block:: python
-
-                mdb.models[name].IncidentWaveProperty
-
-        Parameters
-        ----------
-        name
-            A String specifying the interaction property repository key.
-        definition
-            A SymbolicConstant specifying the type of wave to be defined. Possible values are
-            PLANAR, SPHERICAL, DIFFUSE, AIR_BLAST, and SURFACE_BLAST. The default value is PLANAR.
-        propagationModel
-            A SymbolicConstant specifying the spherical propagation model. Possible values are
-            ACOUSTIC, UNDEX_CHARGE, and GENERALIZED_DECAY. The default value is ACOUSTIC.This
-            argument is valid only when **definition** = SPHERICAL.
-        soundSpeed
-            A Float specifying the speed of sound in the fluid.This argument is not valid when
-            **definition** = AIR_BLAST or when **definition** = SURFACE_BLAST.
-        fluidDensity
-            A Float specifying the fluid mass density.This argument is not valid when
-            **definition** = AIR_BLAST or when **definition** = SURFACE_BLAST.
-        specificHeatRatio
-            None or a Float specifying the ratio of specific heats for gas. The default value is
-            None.This argument is valid only when **definition** = SPHERICAL and
-            **propagationModel** = UNDEX_CHARGE.
-        gravity
-            None or a Float specifying the acceleration due to gravity. The default value is
-            None.This argument is valid only when **definition** = SPHERICAL and
-            **propagationModel** = UNDEX_CHARGE.
-        atmosphericPressure
-            None or a Float specifying the atmospheric pressure. The default value is None.This
-            argument is valid only when **definition** = SPHERICAL and **propagationModel** = UNDEX_CHARGE.
-        dragCoefficient
-            None or a Float specifying the fluid drag coefficient. The default value is None.This
-            argument is valid only when **definition** = SPHERICAL and **propagationModel** = UNDEX_CHARGE.
-        dragExponent
-            A Float specifying the fluid drag exponent. The default value is 2.0.This argument is
-            valid only when **definition** = SPHERICAL and **propagationModel** = UNDEX_CHARGE.
-        waveEffects
-            A Boolean specifying whether or not to include wave effects in the fluid and gas. The
-            default value is ON.This argument is valid only when **definition** = SPHERICAL and
-            **propagationModel** = UNDEX_CHARGE.
-        chargeDensity
-            None or a Float specifying the density of the charge material. The default value is
-            None.This argument is valid only when **definition** = SPHERICAL and
-            **propagationModel** = UNDEX_CHARGE.
-        chargeMass
-            None or a Float specifying the mass of the charge material. The default value is
-            None.This argument is valid only when **definition** = SPHERICAL and
-            **propagationModel** = UNDEX_CHARGE.
-        constantK1
-            None or a Float specifying the charge material constant K. The default value is
-            None.This argument is valid only when **definition** = SPHERICAL and
-            **propagationModel** = UNDEX_CHARGE.
-        constantK2
-            None or a Float specifying the charge material constant k. The default value is
-            None.This argument is valid only when **definition** = SPHERICAL and
-            **propagationModel** = UNDEX_CHARGE.
-        constantA
-            None or a Float specifying the charge material constant A. The default value is
-            None.This argument is valid only when **definition** = SPHERICAL and
-            **propagationModel** = UNDEX_CHARGE.
-        constantB
-            None or a Float specifying the charge material constant B. The default value is
-            None.This argument is valid only when **definition** = SPHERICAL and
-            **propagationModel** = UNDEX_CHARGE.
-        constantKc
-            None or a Float specifying the charge material constant Kc. The default value is
-            None.This argument is valid only when **definition** = SPHERICAL and
-            **propagationModel** = UNDEX_CHARGE.
-        duration
-            None or a Float specifying the time duration for the bubble simulation. The default
-            value is None.This argument is valid only when **definition** = SPHERICAL and
-            **propagationModel** = UNDEX_CHARGE.
-        maximumSteps
-            An Int specifying the maximum number of time steps for the bubble simulation. The
-            default value is 1500.This argument is valid only when **definition** = SPHERICAL and
-            **propagationModel** = UNDEX_CHARGE.
-        relativeStepControl
-            A Float specifying the relative step size control parameter. The default value is
-            1×10-11.This argument is valid only when **definition** = SPHERICAL and
-            **propagationModel** = UNDEX_CHARGE.
-        absoluteStepControl
-            A Float specifying the absolute step size control parameter. The default value is
-            1×10-11.This argument is valid only when **definition** = SPHERICAL and
-            **propagationModel** = UNDEX_CHARGE.
-        stepControlExponent
-            A Float specifying the step size control exponent. The default value is 0.2.This
-            argument is valid only when **definition** = SPHERICAL and **propagationModel** = UNDEX_CHARGE.
-        genDecayA
-            A Float specifying the constant A associated with the generalized decay propagation
-            model. The default value is 0.0.This argument is valid only when **definition** = SPHERICAL
-            and **propagationModel** = GENERALIZED_DECAY.
-        genDecayB
-            A Float specifying the constant B associated with the generalized decay propagation
-            model. The default value is 0.0.This argument is valid only when **definition** = SPHERICAL
-            and **propagationModel** = GENERALIZED_DECAY.
-        genDecayC
-            A Float specifying the constant C associated with the generalized decay propagation
-            model. The default value is 0.0.This argument is valid only when **definition** = SPHERICAL
-            and **propagationModel** = GENERALIZED_DECAY.
-        seedNumber
-            An Int specifying the seed number (N) for the diffuse source calculation. N2 sources
-            will be used in the simulation.This argument is valid only when **definition** = DIFFUSE.
-        massTNT
-            A Float specifying the equivalent mass of TNT, in any preferred mass unit.This argument
-            is valid only when **definition** = AIR_BLAST or **definition** = SURFACE_BLAST.
-        massFactor
-            A Float specifying the multiplication factor to convert from the preferred mass unit to
-            kilograms. The default value is 1.0.This argument is valid only when
-            **definition** = AIR_BLAST or **definition** = SURFACE_BLAST.
-        lengthFactor
-            A Float specifying the multiplication factor to convert from the analysis length unit to
-            meters. The default value is 1.0.This argument is valid only when **definition** = AIR_BLAST
-            or **definition** = SURFACE_BLAST.
-        timeFactor
-            A Float specifying the multiplication factor to convert from the analysis time unit to
-            seconds. The default value is 1.0.This argument is valid only when
-            **definition** = AIR_BLAST or **definition** = SURFACE_BLAST.
-        pressureFactor
-            A Float specifying the multiplication factor to convert from the analysis pressure unit
-            to pascals. The default value is 1.0.This argument is valid only when
-            **definition** = AIR_BLAST or **definition** = SURFACE_BLAST.
-
-        Returns
-        -------
-        IncidentWaveProperty
-            An :py:class:`~abaqus.Interaction.IncidentWaveProperty.IncidentWaveProperty` object.
-        """
-        self.interactions[name] = interaction = IncidentWaveProperty(
-            name,
-            definition,
-            propagationModel,
-            soundSpeed,
-            fluidDensity,
-            specificHeatRatio,
-            gravity,
-            atmosphericPressure,
-            dragCoefficient,
-            dragExponent,
-            waveEffects,
-            chargeDensity,
-            chargeMass,
-            constantK1,
-            constantK2,
-            constantA,
-            constantB,
-            constantKc,
-            duration,
-            maximumSteps,
-            relativeStepControl,
-            absoluteStepControl,
-            stepControlExponent,
-            genDecayA,
-            genDecayB,
-            genDecayC,
-            seedNumber,
-            massTNT,
-            massFactor,
-            lengthFactor,
-            timeFactor,
-            pressureFactor,
         )
         return interaction
 
@@ -2562,275 +1763,6 @@ class InteractionModel(
             thickness,
             smooth,
             contactControls,
-        )
-        return interaction
-
-    def StdContactControl(
-        self,
-        name: str,
-        stiffnessScaleFactor: float = 1,
-        penetrationTolChoice: SymbolicConstant = RELATIVE,
-        relativePenetrationTolerance: float = None,
-        absolutePenetrationTolerance: float = None,
-        frictionOnset: SymbolicConstant = None,
-        automaticTolerances: Boolean = OFF,
-        maxchp: int = 0,
-        perrmx: float = 0,
-        uerrmx: float = 0,
-        stabilizeChoice: SymbolicConstant = NONE,
-        dampFactor: float = 1,
-        dampCoef: float = 0,
-        tangFraction: float = 1,
-        eosFraction: float = 0,
-        zeroDampingChoice: SymbolicConstant = COMPUTE,
-        zeroDamping: float = None,
-        enforceWithLagrangeMultipliers: SymbolicConstant = DEFAULT,
-    ) -> StdContactControl:
-        """This method creates an StdContactControl object.
-
-        .. note:: 
-            This function can be accessed by:
-
-            .. code-block:: python
-
-                mdb.models[name].StdContactControl
-
-        Parameters
-        ----------
-        name
-            A String specifying the contact controls repository key.
-        stiffnessScaleFactor
-            A Float specifying the factor by which Abaqus/Standard will scale the default penalty
-            stiffness to obtain the stiffnesses used for the contact pairs. Only contact
-            interactions defined with augmented Lagrangian surface behavior will be affected by this
-            argument. The default value is 1.0.
-        penetrationTolChoice
-            A SymbolicConstant specifying whether the allowable penetration is an absolute value or
-            a value relative to the characteristic contact surface face dimension. Only contact
-            interactions defined with augmented Lagrangian surface behavior will be affected by this
-            argument. Possible values are RELATIVE and ABSOLUTE. The default value is RELATIVE.
-        relativePenetrationTolerance
-            A Float specifying the ratio of the allowable penetration to the characteristic contact
-            surface face dimension. The float values represent percentages (e.g.: 0.001=0.1%). Only
-            contact interactions defined with augmented Lagrangian surface behavior will be affected
-            by this argument. The default value is 10-3.The **relativePenetrationTolerance** argument
-            applies only when **penetrationTolChoice** = RELATIVE. The **relativePenetrationTolerance**
-            and **absolutePenetrationTolerance** arguments are mutually exclusive.
-        absolutePenetrationTolerance
-            None or a Float specifying the allowable penetration. Only contact interactions defined
-            with augmented Lagrangian surface behavior will be affected by this argument. The
-            **absolutePenetrationTolerance** argument applies only when
-            **penetrationTolChoice** = ABSOLUTE. The **relativePenetrationTolerance** and
-            **absolutePenetrationTolerance** arguments are mutually exclusive. The default value is
-            None.
-        frictionOnset
-            A SymbolicConstant specifying when the application of friction occurs. Possible values
-            are:
-            
-            - IMMEDIATE, specifying the friction is included in the increment when contact occurs.
-            - DELAYED, specifying the application of friction is delayed until the increment after
-              contact occurs.
-        automaticTolerances
-            A Boolean specifying whether Abaqus/Standard should automatically compute an overclosure
-            tolerance and a separation tolerance to prevent chattering in contact. The default value
-            is OFF.The **automaticTolerances** argument cannot be used with the **maxchp**, **perrmx**,
-            and **uerrmx** arguments.
-        maxchp
-            An Int specifying the maximum number of points that are permitted to violate contact
-            conditions in any increment. The default value is 0.Either the **perrmx** or the **uerrmx**
-            argument must be specified in conjunction with the **maxchp** argument.
-        perrmx
-            A Float specifying the maximum value of tensile stress (tensile force in GAP- or
-            ITT-type contact elements) allowed to be transmitted at a contact point. The default
-            value is 0.0.The **perrmx** argument must be specified in conjunction with the **maxchp**
-            argument.
-        uerrmx
-            A Float specifying the maximum overclosure distance allowed at a secondary node that is
-            considered to be open. The default value is 0.0.The **uerrmx** argument must be specified
-            in conjunction with the **maxchp** argument.
-        stabilizeChoice
-            A SymbolicConstant specifying whether or not viscous damping will be specified, and if
-            so, how it will be specified. Possible values are NONE, AUTOMATIC, and COEFFICIENT. The
-            default value is NONE.
-        dampFactor
-            A Float specifying the value of the damping factor. This value is multiplied by the
-            calculated damping coefficient. The default value is 1.0.This argument is only valid
-            when **stabilizeChoice** = AUTOMATIC.
-        dampCoef
-            A Float specifying the damping coefficient. The default value is 0.0.This argument is
-            only valid when **stabilizeChoice** = COEFFICIENT.
-        tangFraction
-            A Float specifying the tangential stabilization as a fraction of the normal
-            stabilization (damping). The default value is 1.0.This argument is valid only if
-            **stabilizeChoice** = AUTOMATIC or COEFFICIENT.
-        eosFraction
-            A Float specifying the fraction of the damping that remains at the end of the step. The
-            default value is 0.0.This argument is valid only if **stabilizeChoice** = AUTOMATIC or
-            COEFFICIENT.
-        zeroDampingChoice
-            A SymbolicConstant specifying how the zero-damping clearance will be specified. Possible
-            values are COMPUTE and SPECIFY. The default value is COMPUTE.This argument is valid only
-            if **stabilizeChoice** = AUTOMATIC or COEFFICIENT.
-        zeroDamping
-            None or a Float specifying the clearance at which damping becomes zero. This argument is
-            valid only when **zeroDampingChoice** = SPECIFY. This argument is valid only if
-            **stabilizeChoice** = AUTOMATIC or COEFFICIENT. The default value is None.
-        enforceWithLagrangeMultipliers
-            A SymbolicConstant specifying whether to enforce the contact constraints with Lagrange
-            multipliers. Possible values are DEFAULT, ENFORCEMENT_OFF, and ENFORCEMENT_ON. The
-            default value is DEFAULT.
-
-        Returns
-        -------
-        StdContactControl
-            A :py:class:`~abaqus.Interaction.StdContactControl.StdContactControl` object.
-
-        Raises
-        ------
-        RangeError
-        """
-        self.interactions[name] = interaction = StdContactControl(
-            name,
-            stiffnessScaleFactor,
-            penetrationTolChoice,
-            relativePenetrationTolerance,
-            absolutePenetrationTolerance,
-            frictionOnset,
-            automaticTolerances,
-            maxchp,
-            perrmx,
-            uerrmx,
-            stabilizeChoice,
-            dampFactor,
-            dampCoef,
-            tangFraction,
-            eosFraction,
-            zeroDampingChoice,
-            zeroDamping,
-            enforceWithLagrangeMultipliers,
-        )
-        return interaction
-
-    def StdInitialization(
-        self,
-        name: str,
-        overclosureType: SymbolicConstant = ADJUST,
-        interferenceDistance: float = None,
-        clearanceDistance: float = None,
-        openingTolerance: float = None,
-        overclosureTolerance: float = None,
-    ) -> StdInitialization:
-        """This method creates a StdInitialization object.
-
-        .. note:: 
-            This function can be accessed by:
-
-            .. code-block:: python
-
-                mdb.models[name].StdInitialization
-
-        Parameters
-        ----------
-        name
-            A String specifying the contact initialization repository key.
-        overclosureType
-            A SymbolicConstant specifying the type of overclosure to be defined. Possible values are
-            ADJUST, INTERFERENCE, and CLEARANCE. The default value is ADJUST.
-        interferenceDistance
-            None or a Float specifying the interference distance. This argument is valid only when
-            **overclosureType** = INTERFERENCE. The default value is None.
-        clearanceDistance
-            None or a Float specifying the initial clearance distance. This argument is valid only
-            when **overclosureType** = CLEARANCE, and must be specified in that case. The default value
-            is None.
-        openingTolerance
-            None or a Float specifying the distance tolerance within which initial openings will
-            undergo strain-free adjustments. This argument is not valid when
-            **overclosureType** = INTERFERENCE unless a value has been specified for
-            **interferenceDistance**. The default value is None.
-        overclosureTolerance
-            None or a Float specifying the distance tolerance within which initial overclosures will
-            undergo strain-free adjustments.. The default value is None.
-
-        Returns
-        -------
-        StdInitialization
-            A :py:class:`~abaqus.Interaction.StdInitialization.StdInitialization` object.
-
-        Raises
-        ------
-        RangeError
-        """
-        self.interactions[name] = interaction = StdInitialization(
-            name,
-            overclosureType,
-            interferenceDistance,
-            clearanceDistance,
-            openingTolerance,
-            overclosureTolerance,
-        )
-        return interaction
-
-    def StdStabilization(
-        self,
-        name: str,
-        zeroDistance: float = None,
-        reductionFactor: float = 0,
-        scaleFactor: float = 1,
-        tangentialFactor: float = 0,
-        amplitude: str = "",
-        reset: Boolean = OFF,
-    ) -> StdStabilization:
-        """This method creates a StdStabilization object.
-
-        .. note:: 
-            This function can be accessed by:
-
-            .. code-block:: python
-
-                mdb.models[name].StdStabilization
-
-        Parameters
-        ----------
-        name
-            A String specifying the contact stabilization repository key.
-        zeroDistance
-            None or a Float specifying the clearance distance at which the stabilization becomes
-            zero. The default value is None.
-        reductionFactor
-            A Float specifying the factor by which the analysis will reduce the contact
-            stabilization coefficient per increment. The default value is 0.1.
-        scaleFactor
-            A Float specifying the factor by which the analysis will scale the contact stabilization
-            coefficient. The default value is 1.0.
-        tangentialFactor
-            A Float specifying the factor that scales the contact stabilization coefficient in the
-            tangential direction. The default value is 0.0.
-        amplitude
-            A String specifying the name of the Amplitude object that defines a time-dependent scale
-            factor for contact stabilization over the step. The default value is an empty string.
-        reset
-            A Boolean specifying whether to cancel carryover effects from contact stabilization
-            specifications involving nondefault amplitudes that appeared in previous steps. The
-            default value is OFF.
-
-        Returns
-        -------
-        StdStabilization
-            A :py:class:`~abaqus.Interaction.StdStabilization.StdStabilization` object.
-
-        Raises
-        ------
-        RangeError
-        """
-        self.interactions[name] = interaction = StdStabilization(
-            name,
-            zeroDistance,
-            reductionFactor,
-            scaleFactor,
-            tangentialFactor,
-            amplitude,
-            reset,
         )
         return interaction
